skills/parity-mapping/scripts/requirements.txt
requests>=2.28
pyyaml>=6.0
rapidfuzz>=3.0
pandas>=2.0
networkx>=3.0
